pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.3.0",
    "jsonschema>=4.23.0",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "packaging>=24.2",
    "pydantic>=2.11.7",
    "pygithub>=2.6.1",
    "typing-extensions>=4.14.0",
    "watchdog>=6.0.0",
    "google-generativeai>=0.8.3",
    "anthropic>=0.39.0",
    "openai>=1.58.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
pr-agents = "pr_agents.cli:main"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    "tests/fixtures/prebid_scenarios.py",  # Contains intentionally misformatted test data
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  # test fixtures with intentionally misformatted data
  | tests/fixtures/prebid_scenarios\.py
)/
'''
